+++ CHANGELOG.md
@@ -1,11 +1,67 @@
+#### v1.1.1 (2017-01-17)
+
+- 【修复】 fix icon-safe的错乱问题
+- 【增强】 增加按钮loading状态
+- 【增强】 更新普通警告图标颜色
+
+
+
+#### v1.1.0 (2016-11-21)
+
+- 【增强】 新增兼容IE Edge版本的switch
+- 【增强】 增加组件 badge徽章
+- 【增强】 增加组件 slider滑块
+- 【增强】 增加组件 picker多列选择器
+- 【优化】 把radio和checkbox下的weui-cell的:active挪到weui-check__label
+- 【优化】 src/navbar.less 更名为 src/weui-navbar.less
+- 【优化】 src/tabbar.less 更名为 src/weui-tabbar.less
+- 【优化】 loading的base64
+- 【优化】 去掉废弃和冗余的样式
+- 【修复】 preview下，两个button并排时，高度有问题 (#462)
+- 【修复】 button作为.weui-vcode-btn时的样式问题 (#519)
+- 【修复】 sourcemap 为乱码的问题
+- 【修复】 select的箭头没有居中的问题
+- 【修复】 tabbar的间距样式
+
+
+
+#### v1.0.2 (2016-09-27)
+
+- 【修复】 fix 不见了的weui-icon_clear的icon: weui-icon_close -> weui-icon_clear
+- 【修复】 preview在没有value的时候会变形: 增加weui-form-preview__item来控制
+- 【修复】 补上九宫格Grid的DEMO
+- 【修复】 九宫格文字换行问题 #483
+- 【优化】 icon的实现
+- 【优化】 example代码
+
+
+#### v1.0.1 (2016-09-27)
+
+- 【修复】 preview中，button与a标签的样式不一致
+- 【修复】 把遗漏的weui-media 更名为 weui-media-box
+
+
+#### v1.0.0 (2016-09-23)
+
+- 【增强】 采用BEM命名规范
+- 【增强】 新增Flex布局
+- 【增强】 新增Gallery
+- 【增强】 新增Preview
+- 【增强】 新增Agreement
+- 【增强】 新增Footer
+- 【增强】 新增Dialog和Actionsheet的Android样式
+- 【修复】 修正了部分组件的间距
+- 【修复】 修复了已知问题
+
+
 #### v0.4.3 (2016-07-05)
 
-- [修复] 图片上传组件无法选中图片的缺陷
-- [修复] 九宫格组件少于3个时出现多余线段的缺陷
-- [修复] 解决某些组件的z-index冲突，标准化z-index
-- [修复] 解决`pannel`图标没有对齐的问题 
-- [增强] demo 更新 router 
-- [增强] autoprefixer 指定浏览器范围
+- 【修复】 图片上传组件无法选中图片的缺陷
+- 【修复】 九宫格组件少于3个时出现多余线段的缺陷
+- 【修复】 解决某些组件的z-index冲突，标准化z-index
+- 【修复】 解决`pannel`图标没有对齐的问题 
+- 【增强】 demo 更新 router 
+- 【增强】 autoprefixer 指定浏览器范围
 
 详情参考:
 
@@ -18,12 +74,12 @@
 
 #### v0.4.2 (2016-04-28)
 
-- [修复] 修复 `weui_label` 在英文字符下溢出的缺陷
-- [增强] Article 页面增加支持图片样式
-- [修复] demo 页面的 `lang` 属性修改
-- [修复] 修复 demo 页面的 ActionSheet mask 无动画的缺陷
-- [增强] 优化 demo, 为tab中navbar和tabbar里标签添加点击事件交互，方便查看实际效果
-- [修复] 修改 `weui_label` 的宽度为 105px
+- 【修复】 修复 `weui_label` 在英文字符下溢出的缺陷
+- 【增强】 Article 页面增加支持图片样式
+- 【修复】 demo 页面的 `lang` 属性修改
+- 【修复】 修复 demo 页面的 ActionSheet mask 无动画的缺陷
+- 【增强】 优化 demo, 为tab中navbar和tabbar里标签添加点击事件交互，方便查看实际效果
+- 【修复】 修改 `weui_label` 的宽度为 105px
 
 详情参考: